cooker: Fix/improve collections handling
Code changes for FILE_LAYERNAME exposed lifecycle issues around the collections
objection in Cooker which only appeared in devtool usage in eSDK. Move
the collections setup to an earlier stage after parsing compeltes to
avoid any kind of race around it.

Also stop overwriting the code variable in MatchFiles. Ultimately we
need to combine these codepaths but that is for another patch.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>